# feat(scope-manager): only populate implicit globals whose names appear in the file