## PR Checklist

- [x] Addresses an existing open issue: fixes #9575
- [x] That issue was marked as [accepting prs](https://github.com/typescript-eslint/typescript-eslint/issues?q=is%3Aopen+is%3Aissue+label%3A%22accepting+prs%22)
- [x] Steps in [Contributing](https://typescript-eslint.io/contributing) were taken

## Overview

This PR optimizes `populateGlobalsFromLib` by deduplicating work to calculate the set of variables available in the configured `libs` array, and by only injecting implicit variable definitions for variables whose names appear in the file being analyzed.

As a consequence of only injecting implicit variables that are actually used in the file, we can avoid the overhead of lint rules spending time processing irrelevant injected globals, for example in `typescript-eslint/no-redeclare` and `typescript-eslint/naming-convention`.

## Before
![populateGlobalsFromLib took 8.556 seconds](https://github.com/user-attachments/assets/288a17a0-f031-4632-b426-717ba586f264)
![findVariablesInScope took 4.077 seconds](https://github.com/user-attachments/assets/71644883-2484-425b-8851-e8ac75b7625d)

## After
![populateGlobalsFromLib took 0.312 seconds](https://github.com/user-attachments/assets/e680485b-d978-49bc-a334-da2235f4ff96)
![findVariablesInScope took 0.662 seconds](https://github.com/user-attachments/assets/63730200-72fe-4617-878a-64ced4f5d7be)