@bradzacher 2024-12-27
As shown in the failed tests this doesn't work as written. I'm guessing that what's happening is that you're deferring the addition of the variables to after the scopes are closed - causing dangingling, unresolved references.

When a scope is finalised we close off the references within the scope. This means we attempt to resolve any unresolved references created in the current or child scopes to variables defined in the current scope. If we don't find the right variable the reference is left as unresolved and we continue to the parent scope.

So if you don't declare the implicit globals early enough then they can't be resolved and so the references are left dangling - which then causes the  `no-undef` rule to error on them.