@JoshuaKGoldberg 2025-01-13
This is really cool, thanks for sending @dmichon-msft! I'd love to dive in soon and see if we can get this shipped. But - we're a little swamped catching up on issues after the holiday & new year break, and this is a lot of very nuanced code in a not-very-often changed area of the project. So it'll probably be a little while till we can take a deep look.

In the meantime, what would help us review more quickly:

* I see there are some _"Added ... was not covered by tests"_ annotations: could you either test those, or leave comments explaining that they are tested but the annotations are inaccurate (e.g. from #6701 or similar)?
* The `populateGlobalsFromLib` function has gotten huge and IMO much less readable. maybe split out some well-named helper functions to encapsulate more of the used-only-in-one-place logic/variables?
    * Example: `entriesForLib` could be a `const` if the retrieval logic were moved to a separate function
    * Example: `implicitVariablesFromLibs`'s creation & cache checking could be moved to a separate function

Btw, you don't need to keep merging the latest `main` in: https://typescript-eslint.io/contributing/pull-requests#updating-over-time. We don't want to give that burden to contributors. Especially if us taking a while to get to a PR makes it worse!