@dmichon-msft 2025-01-13
I'm slicing out the easy perf win (caching of `Object.entries` calls) into its own PR that will be much simpler to review, then I'll refactor this a bit and add more testing to the pieces that power it.