@JoshuaKGoldberg 2025-02-24
> I'm slicing out the easy perf win (caching of `Object.entries` calls) into its own PR that will be much simpler to review, then I'll refactor this a bit and add more testing to the pieces that power it.

👍 in that case drafting this PR in the meantime. Let me know if it's actually ready for review and I misinterpreted. Exciting stuff! 🙂