@JoshuaKGoldberg 2025-04-07
👋 Just checking in @dmichon-msft, do you still have time for this?