@JoshuaKGoldberg 2025-06-30
I don't mean to bug you @dmichon-msft but I think it's a little unclear - do you have an ETA for when you'd be able to get back to this? If so, that'd be wonderful, we'd love to have this change merged. If not, we'd prefer to close the PR to let any other interested community members pitch in. WDYT?